Add if statement.

+++ install-scripts/2-baseSystem/04-timeServer.sh
@@ -3,7 +3,9 @@
 mkdir -p /var/log/chrony
 
 apt-get install -y -q chrony at
-systemctl disable systemd-timesyncd || true
+if service --status-all | grep -Fq 'systemd-timesyncd'; then
+  systemctl disable systemd-timesyncd || true
+fi
 
 ## TimeZone
 ln -fs /usr/share/zoneinfo/Etc/UTC /etc/localtime
